refactor: 멘토링 피드백 및 리팩토링
