+++ frontend/src/App.css
@@ -1,3 +1,7 @@
+body {
+  box-sizing: border-box;
+}
+
 #topbar {
   display: flex;
   justify-content: space-between;
@@ -51,51 +55,6 @@ p {
   margin-right: 100px;
 }
 
-#clickAvailable {
-  cursor: pointer;
-}
-
-#foodtype {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  margin: 0 auto;
-}
-
-#foodtype > div {
-  margin: 10px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-decoration: none;
-  border: darkgreen solid 1px;
-  width: 120px;
-  height: 50px;
-  border-radius: 10px;
-  padding-top: 10px;
-  background-color: #eee;
-  margin-top: 30px;
-  font-weight: bold;
-}
-#foodtype div.selected {
-  color: darkgreen;
-  border: 2px solid green;
-}
-
-#foodtype > div:hover {
-  background-color: #fff;
-  color: darkgreen;
-  border: 2px solid green;
-}
-
-#foodtype img {
-  width: 20px;
-  height: 20px;
-  margin-bottom: 5px;
-  flex-direction: column;
-  align-items: center;
-}
-
 #place {
   font-size: 12px;
   display: flex;
@@ -119,6 +78,7 @@ p {
   font-size: 13px;
   width: 120px;
   height: 30px;
+  cursor: pointer;
 }
 
 #recruit-button {
