+++ frontend/src/components/auth/index.js
@@ -1,0 +1,2 @@
+export * from './Login';
+export * from './Register';
