+++ frontend/src/components/detail/index.js
@@ -1,2 +1,4 @@
-export { PostDetail } from './PostDetail';
 export { CommentForm } from './CommentForm';
+export { PostContents } from './PostContents';
+export { PostDetail } from './PostDetail';
+export { StatusList } from './StatusList';
