+++ frontend/src/components/index.js
@@ -1,3 +1,6 @@
+export * from './auth';
 export * from './detail';
-export * from './login';
+export * from './header';
 export * from './main';
+export * from './menu';
+export * from './posts';
