--- frontend/src/components/login/index.js
@@ -1,3 +1,0 @@
-// eslint-disable-next-line import/named, import/prefer-default-export
-export * from './Login';
-export * from './Register';
