+++ frontend/src/components/main/Main.jsx
@@ -1,0 +1,11 @@
+import { Header } from '../header';
+import { Posts } from '../posts';
+
+export function Main() {
+  return (
+    <>
+      <Header />
+      <Posts />
+    </>
+  );
+}
