+++ frontend/src/components/main/index.js
@@ -1,4 +1,1 @@
 export * from './Main';
-export * from './Logo';
-export * from './Posts';
-export * from './TopBar';
