+++ frontend/src/components/main/main.const.js
@@ -1,28 +1,1 @@
-export const FOOD_TYPE_LIST = [
-  {
-    id: 'western',
-    food: '양식',
-  },
-  {
-    id: 'franchise',
-    food: '프랜차이즈',
-  },
-  {
-    id: 'korean',
-    food: '한식',
-  },
-  {
-    id: 'japanese',
-    food: '일식',
-  },
-  {
-    id: 'vegan',
-    food: '채식',
-  },
-  {
-    id: 'dessert',
-    food: '디저트',
-  },
-];
-
 export const BUILDING_LIST = ['한우리집', '이하우스', '아이하우스', '학생문화관', '중앙도서관', '교외'];
