+++ frontend/src/components/menu/Menu.jsx
@@ -1,14 +1,9 @@
 import { useState } from 'react';
-import { FOOD_TYPE_LIST } from './main.const';
+import { FoodTypeItem } from './FoodTypeItem';
+import { FOOD_TYPE_LIST } from './menu.const';
+import * as S from './menu.style';
 
-const FoodTypeItem = ({ id, food, selected, onFoodTypeClick }) => (
-  <div onClick={() => onFoodTypeClick(id)} className={selected ? 'selected' : ''} id="clickAvailable">
-    <img src={`/assets/${id}-icon.svg`} alt={`/${food}`} />
-    <p>{food}</p>
-  </div>
-);
-
-const FoodTypes = ({ onFoodtypeChange }) => {
+export const Menu = ({ onFoodTypeChange }) => {
   const [selectedFoodType, setSelectedFoodType] = useState('');
 
   const handleFoodTypeClick = clickedId => {
@@ -16,11 +11,11 @@ const FoodTypes = ({ onFoodtypeChange }) => {
     setSelectedFoodType(prevSelected => (prevSelected === clickedId ? '' : clickedId));
 
     // 선택된 음식 카테고리를 상위 컴포넌트로 전달
-    onFoodtypeChange(prevSelected => (prevSelected === clickedId ? '' : clickedId));
+    onFoodTypeChange(prevSelected => (prevSelected === clickedId ? '' : clickedId));
   };
 
   return (
-    <div id="foodtype">
+    <S.Menu>
       {FOOD_TYPE_LIST.map(({ id, food }) => (
         <FoodTypeItem
           key={id}
@@ -30,8 +25,6 @@ const FoodTypes = ({ onFoodtypeChange }) => {
           onFoodTypeClick={handleFoodTypeClick}
         />
       ))}
-    </div>
+    </S.Menu>
   );
 };
-
-export default FoodTypes;
