+++ frontend/src/components/menu/PlaceDropdown.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
-import { BUILDING_LIST } from './main.const';
+import { BUILDING_LIST } from '../main/main.const';
 
 // TODO: window.location.href switch문 삭제 ('Posts.js'에 필터링 함수 추가):
-export default function DropdownExample({ onBuildingChange }) {
+export function PlaceDropdown({ onBuildingChange }) {
   const [selectedBuilding, setSelectedBuilding] = useState('');
 
   const handleBuildingChange = e => {
