+++ frontend/src/components/menu/index.js
@@ -1,0 +1,4 @@
+export * from './Menu';
+export * from './PlaceDropdown';
+
+export * from './menu.const';
