+++ frontend/src/components/menu/menu.const.js
@@ -1,0 +1,26 @@
+export const FOOD_TYPE_LIST = [
+  {
+    id: 'western',
+    food: '양식',
+  },
+  {
+    id: 'franchise',
+    food: '프랜차이즈',
+  },
+  {
+    id: 'korean',
+    food: '한식',
+  },
+  {
+    id: 'japanese',
+    food: '일식',
+  },
+  {
+    id: 'vegan',
+    food: '채식',
+  },
+  {
+    id: 'dessert',
+    food: '디저트',
+  },
+];
