+++ frontend/src/components/posts/index.js
@@ -1,0 +1,2 @@
+export * from './Posts';
+export * from './posts.const';
